refactor: extract pane terminal helpers
